Extract variable menu to its own component
