+++ packages/slate-editor/src/extensions/variables/VariableMenu.module.scss
@@ -1,0 +1,22 @@
+@import "styles/helpers";
+@import "styles/variables";
+
+.Dropdown {
+    width: 100% !important;
+}
+
+.Caption {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    .Icon {
+        display: flex;
+        align-items: center;
+
+        > svg {
+            width: 16px;
+            height: 16px;
+        }
+    }
+}
